Minimum Rust structure
This is for study purposes, as this whole repo has always been. The
idea is to replace parts of the Python code with Rust implementation
for performance gains — and maybe to simplify the stack (maybe remove
the database, for example).

By now what is needed:

```
$ poetry install # make sure maturin is available on the venv
$ poetry run pip install . # make sure crates are loaded in the venv
```

Example of use and benchmark:

```
$ python -m timeit -n 1000 -s "from crates import latest_changed_at" "latest_changed_at('.')"
1000 loops, best of 5: 18.4 msec per loop
$ python -m timeit -n 1000 -s "from whiskyton.helpers.sitemap import most_recent_update" "most_recent_update('.')"
1000 loops, best of 5: 118 msec per loop``
```